requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
databases==0.8.0
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
importlib-metadata==6.7.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
